+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 This document contains all changes to the crate since version 0.1.8.
 
+# 0.3.0 (unreleased)
+
+- Made all quadrature rules store their nodes sorted in ascending order.
+ This means that all functions that return some view of the nodes (and weights) now return them in sorted order.
+ This affects the Gauss-Legendre, Gauss-Hermite and Gauss-Chebyshev rules.
+ The affected functions are `QuadratureRule::iter()`, `QuadratureRule::into_iter()`, `QuadratureRule::nodes()`, `QuadratureRule::weights()`, `QuadratureRule::as_node_weight_pairs()` and `QuadratureRule::into_node_weight_pairs()`.
+ 
+
 # 0.2.2
 
 - Add Gauss-Chebyshev quadrature of the first and second kinds.